refactor: 경매 상세 정보 조회 API 전환 (#123)
* refactor: 좋아요수, 입찰 수 Long 타입으로 변경

* feat: 경매 상세 조회 API 추가

* feat: 경매 상태를 찾는 쿼리 추가

* feat: 경매 상세정보 조회 쿼리 추가

* chore: Spring security API 인증 허용

* feat: 경매 상세 정보 조회 추상클래스 추가

* feat: 사전 경매 상세 조회 응답 객체 추가

* feat: 정식 경매 상세정보 조회 응답 객체 추가

* feat: 전환을 위한 임시 V2 Entity 추가

* feat: 전환을 위한 임시 V2 Entity 스크립트 추가

* feat: 경매 상세 정보 조회 서비스 함수 추가

* feat: V2 Payment Method 추가

* feat: V2 Payment Method 추가

* test: 경매 상세정보 조회 테스트 추가

* refactor: 명확한 클래스명으로 수정

- AuctionDetailBaseResponse -> BaseAuctionDetailResponse

+++ src/main/java/org/chzz/market/domain/auctionv2/entity/AuctionV2.java
@@ -79,11 +79,11 @@ public class AuctionV2 extends BaseTimeEntity {
 
     @Builder.Default
     @Column
-    private Integer likeCount = 0;
+    private Long likeCount = 0L;
 
     @Builder.Default
     @Column
-    private Integer bidCount = 0;
+    private Long bidCount = 0L;
 
     @Builder.Default
     @OneToMany(mappedBy = "auction", cascade = {CascadeType.REMOVE, CascadeType.PERSIST}, orphanRemoval = true)
